create stats module
